+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-cash.spec.ts
@@ -77,6 +77,9 @@ describe('PortfolioCalculator', () => {
   let redisCacheService: RedisCacheService;
 
   beforeEach(() => {
+    PortfolioSnapshotServiceMock.reset();
+    RedisCacheServiceMock.reset();
+
     configurationService = new ConfigurationService();
 
     exchangeRateDataService = new ExchangeRateDataService(